backend/requirements.txt
# LiveKit
livekit>=1.0.0
livekit-agents>=1.0.0
livekit-api>=1.0.0

# FastAPI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# HTTP client
httpx>=0.27.0

# Google Generative AI (Gemini)
google-generativeai>=0.8.0

# Audio processing
numpy>=1.26.0
scipy>=1.11.0

# Environment variables
python-dotenv>=1.0.0

# Pydantic (required by FastAPI)
pydantic>=2.0.0

# Visualization generation
manim>=0.19.0
